correção feita ao inserir cartão de crédito para pagamento de um plano de conta

+++ ClassLibrary/Database Model/IUserData.cs
@@ -10,7 +10,7 @@
         Task InsertUser(User user);
         Task PlanFinished(int usernameId);
         List<CreditCard> GetCreditCard(CreditCard card);
-        void Payment(CreditCard card, double cost, string plan, User user, string date);
+        void Payment(CreditCard card, decimal cost, string plan, User user, string date);
         void InsertCreditCard(CreditCard card);
         List<User> GetUserSync(int usernameId);
     }
